server/requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.27
pydantic>=2.6
pandas>=2.2
openai>=1.10
python-multipart
python-dotenv
pdfplumber==0.10.3